Hoist PDF-to-screenshots conversion into shared test-utils helper
Both the scan and mark integration tests render generated PDFs to numbered
PNGs for the screenshot gallery, and more apps will follow. Extract the
common render/number/blank-skip logic into capturePdfScreenshots in
libs/test-utils (next to the existing screenshot helpers), and have each app's
support files supply the PDF bytes from their own source (mock printer, mock
USB drive). Moves the @votingworks/image-utils dependency to test-utils.
