+++ apps/mark/integration-testing/package.json
@@ -19,7 +19,6 @@
     "@votingworks/basics": "workspace:*",
     "@votingworks/fixtures": "workspace:*",
     "@votingworks/grout": "workspace:*",
-    "@votingworks/image-utils": "workspace:*",
     "@votingworks/printing": "workspace:*",
     "@votingworks/test-utils": "workspace:*",
     "@votingworks/types": "workspace:*",
