+++ apps/scan/integration-testing/e2e/support/print_to_png.ts
@@ -1,7 +1,9 @@
 import { readFileSync } from 'node:fs';
 import { getMockFileFujitsuPrinterHandler } from '@votingworks/fujitsu-thermal-printer';
-import { pdfToImages, writeImageData } from '@votingworks/image-utils';
-import type { ScreenshotCounter } from '@votingworks/test-utils';
+import {
+  capturePdfScreenshots,
+  type ScreenshotCounter,
+} from '@votingworks/test-utils';
 
 /** Reads the most recently printed PDF and saves each page as a PNG in test-results/screenshots. */
 export async function capturePrintedReport(
@@ -11,13 +13,5 @@ export async function capturePrintedReport(
   const printPath = getMockFileFujitsuPrinterHandler().getLastPrintPath();
   if (!printPath) return;
   const pdfBytes = new Uint8Array(readFileSync(printPath));
-  let pageIndex = 0;
-  for await (const { page } of pdfToImages(pdfBytes, { scale: 2 })) {
-    const suffix = pageIndex === 0 ? name : `${name}-page${pageIndex + 1}`;
-    await writeImageData(
-      `./test-results/screenshots/${counter.next()}-${suffix}.png`,
-      page
-    );
-    pageIndex += 1;
-  }
+  await capturePdfScreenshots(pdfBytes, name, counter);
 }
