+++ apps/scan/integration-testing/package.json
@@ -20,7 +20,6 @@
     "@votingworks/fujitsu-thermal-printer": "workspace:*",
     "@votingworks/grout": "workspace:*",
     "@votingworks/hmpb": "workspace:*",
-    "@votingworks/image-utils": "workspace:*",
     "@votingworks/pdi-scanner": "workspace:*",
     "@votingworks/test-utils": "workspace:*",
     "@votingworks/types": "workspace:*",
